+++ src/main/java/ua/edu/ukma/ykrukovska/practice2/storage/Storage.java
@@ -3,8 +3,8 @@ package ua.edu.ukma.ykrukovska.practice2.storage;
 public interface Storage {
 
     double getProductAmount(String productName);
-    void addProduct(String productName, double amount);
-    void withdrawProduct(String productName, double amount);
+    void addProduct(String productName, int amount);
+    void withdrawProduct(String productName, int amount);
     void addProductGroup(String productGroupName);
     void addProductToGroup(String productGroupName, String productName);
     void setProductPrice(String productName, double price);
